Manages the aesthetics of the graphics

Manages the shape, color and line type of the graphics according to their content

aesthetics(
  sim,
  obs = NULL,
  type = c("dynamic", "scatter"),
  overlap = NULL,
  several_sit = FALSE,
  shape_sit = c("none", "txt", "symbol", "group"),
  one_version = TRUE,
  first_sim = TRUE
)

Arguments

sim
A simulation list of data.frames

obs
An observation list of data.frames

type
The type of plot required, either "dynamic" or "scatter"

overlap
A list of lists containing the variables to represent on the same graph when type = "dynamic"

several_sit
Boolean. Must be equal to TRUE if sim and obs gather more than one situation and if situations should be differentiated on the graph.

shape_sit
Shape to differentiate between situations when all_situations= TRUE. See details.

one_version
Boolean. Must be TRUE if several version will be plotted on the same graph.

first_sim
Boolean. Must be TRUE is sim corresponds to the first version.

Value

An aesthetics expression which will allow to manage the aesthetics of the graphics

Details

The shape_sit argument can be:

"none" (the default): Same shape for all situations.

"txt": Writes the name of the situation above each point.

"symbol": One shape for each situation.

"group": One shape for each group of situations described in situation_group.