Format a STICS simulation (and observation)

Format a STICS simulation (and observation if any) for plotting. This function can be used as a template to include other models in CroPlotR.

format_stics(sim, obs = NULL, plot = c("sim", "common", "obs", "all"))

Arguments

sim
A simulation list of data.frames

obs
An observation list of data.frames

plot
The priority to either simulation or observation points if missing values (see details)

Value

A list of two: a pre-formatted data.frame, and a colouring expression.

Details

The plot argument can be:

"sim": all variables with simulations outputs, and observations when there are some

"common": variables with simulations outputs and observations in common

"obs": all variables with obervations, and simulations outputs when there are some

"all": all variables with any obervations or simulations outputs

Note

This function is exported only so developers can see the output required for other models.

Examples

workspace= system.file(file.path("extdata", "stics_example_1"), package = "CroPlotR")
situation= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))[1]
sim= SticsRFiles::get_daily_results(workspace = workspace, usm_name = situation)
obs= SticsRFiles::get_obs(workspace =  workspace, usm_name = situation)
formated_df= format_stics(sim$`IC_Wheat_Pea_2005-2006_N0`,obs$`IC_Wheat_Pea_2005-2006_N0`)
options(max.print= 100)
formated_df

#> $df
#>          Date Plant Dominance variable Simulated Observed
#> 1  2005-09-26   ble Principal    lai_n         0       NA
#> 2  2005-09-27   ble Principal    lai_n         0       NA
#> 3  2005-09-28   ble Principal    lai_n         0       NA
#> 4  2005-09-29   ble Principal    lai_n         0       NA
#> 5  2005-09-30   ble Principal    lai_n         0       NA
#> 6  2005-10-01   ble Principal    lai_n         0       NA
#> 7  2005-10-02   ble Principal    lai_n         0       NA
#> 8  2005-10-03   ble Principal    lai_n         0       NA
#> 9  2005-10-04   ble Principal    lai_n         0       NA
#> 10 2005-10-05   ble Principal    lai_n         0       NA
#> 11 2005-10-06   ble Principal    lai_n         0       NA
#> 12 2005-10-07   ble Principal    lai_n         0       NA
#> 13 2005-10-08   ble Principal    lai_n         0       NA
#> 14 2005-10-09   ble Principal    lai_n         0       NA
#> 15 2005-10-10   ble Principal    lai_n         0       NA
#> 16 2005-10-11   ble Principal    lai_n         0       NA
#>  [ reached 'max' / getOption("max.print") -- omitted 1233 rows ]
#> 
#> $coloring
#> $coloring$Plant
#> paste(.data$Dominance, ":", .data$Plant)
#> 
#>