Generic formatting

This function performs generic formatting on a data frame based on the input parameters.

generic_formatting(
  df,
  var,
  overlap,
  situation_group,
  type,
  shape_sit,
  several_sit,
  total_vers,
  num_vers
)

Arguments

df

A data frame to be formatted (e.g. output from format_cropr).

var

A character string indicating the variable to be formatted.

overlap

A logical value indicating whether there is any overlapping variables desired in the plot.

situation_group

A character string indicating the situation group.

type

A character string indicating the type of plot (dynamic or scatter).

shape_sit

A logical value indicating whether the shape should be based on the situation.

several_sit

A logical value indicating whether there are several situations to plot.

total_vers

An integer indicating the total number of versions.

num_vers

An integer indicating the id of the current version.

Value

A formatted data frame.