Parse plot situations arguments

This function parses the arguments for plot_situations.

parse_plot_situations_args(
  dot_args,
  obs = NULL,
  obs_sd = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  var = NULL,
  title = NULL,
  all_situations = TRUE,
  overlap = NULL,
  successive = NULL,
  shape_sit = c("none", "txt", "symbol", "group"),
  situation_group = NULL,
  reference_var = NULL,
  force = FALSE,
  verbose = TRUE
)

Arguments

obs

A list (each element= situation) of observations data.frames (named by situation)

obs_sd

A list (each element= situation) of standard deviations of observations data.frames (named by situation)

type

The type of plot requested, either "dynamic" (date in X, variable in Y) or scatter (simulated VS observed)

select_dyn

Which data to plot when type= "dynamic"? See details.

select_scat

Which data to plot when type= "scatter"? See details.

var

A vector of variables that should be displayed on the graph.

title

A vector of plot titles, named by situation. Use the situation name if NULL, recycled if length one.

all_situations

Boolean (default = TRUE). If TRUE, plot all situations on the same graph.

overlap

A list of lists containing the variables to represent on the same graph when type = "dynamic".

successive

A list of lists containing the situations to be represented as a contiguous sequence when type = "dynamic" (implies that the situations are correctly ordered).

shape_sit

Shape to differentiate between situations when all_situations= TRUE. See details.

situation_group

A list of lists of situations to gather when shape_sit= "group".

reference_var

Variable selected on x-axis when type is scatter and select_scat is res. It is possible to select between observation and simulation of the reference variable. (examples : reference_var = "lai_n_obs", reference_var = "mafruit_sim")

force

Continue if the plot is not possible ? E.g. no observations for scatter plots. If TRUE, return NULL, else return an error (default).

verbose

Boolean. Print information during execution.

Value

A list of parsed arguments, eventually corrected according to the compatibility between options.