Generic ploting of a situation

Plots outputs of a model (and observations) for one situation. This function is used as a generic ploting function for any models. To use it with your own model, please provide a wrapper function around your model to format the outputs used by this function (see format_stics() for a template), and then provide your custom function as an argument to this one.

plot_generic_situation(
  sim,
  obs = NULL,
  type = c("dynamic", "scatter"),
  plot = c("sim", "common", "obs", "all"),
  title = NULL,
  force = TRUE,
  verbose = TRUE,
  formater
)

Arguments

sim
A simulation data.frame

obs
An observation data.frame (variable names must match)

type
The type of plot required, either "dynamic" or "scatter"

plot
The priority to either simulation or observation points if missing values (see details)

title
The plot title

force
Continue if the plot is not possible ? E.g. no observations for scatter plots. If TRUE, return NULL, else return an error.

verbose
Boolean. Print informations during execution.

formater
The function used to format the models outputs and observations in a standard way. You can design your own function that format one situation and provide it here.

Value

A ggplot object

Details

The plot argument can be:

"sim": all variables with simulations outputs, and observations when there are some

"common": variables with simulations outputs and observations in common (this is forced when type="scatter")

"obs": all variables with obervations, and simulations outputs when there are some

"all": all variables with any obervations or simulations outputs