Save CroPlotR plots

Save the plots to a pdf file

save_plot_pdf(
  plot,
  out_dir,
  file_name = "Graphs",
  title = "Plots",
  file_per_var = FALSE,
  stats = NULL,
  force = TRUE,
  verbose = TRUE,
  path = lifecycle::deprecated(),
  filename = lifecycle::deprecated(),
  main = lifecycle::deprecated()
)

Arguments

plot

A list of ggplots : output of plot()

out_dir

The path to the directory where to save the plots

file_name

Name of the pdf file

title

Main title of the pdf document

file_per_var

If TRUE, produces one file per variable instead of one with all variables inside

stats

Output of statistics_situationswith all_situations = FALSE. Needed if file_per_var is TRUE. It is used to classify situations according to the descending RMSE.

force

Continue if the plot is not possible ? If TRUE, return NULL, else return an error.

verbose

Logical value for displaying information while running

path

path is no longer supported, use out_dir instead.

filename

filename is no longer supported, use file_name instead.

main

title is no longer supported, use out_dir instead.

Value

Save the plots in a pdf file in the folder specified by the path