Specific functions to generate dynamic plots

Generate dynamic plots for the different cases handled in CroPlotR (plant mixture, plot several simulation results on same graph, ...) as specified by the different arguments.

plot_dynamic(df_data, sit, successive, title = NULL)

Arguments

df_data

A named list of data frame including the data to plot (one df per situation)

sit

The name of the situation to plot

successive

A list of lists containing the situations to be represented as a contiguous sequence

title

The plot title (optional, NULL by default)

Value

A ggplot object

A list of ggplot objects

Details

List of the different specific functions:

plot_dynamic_mixture: Generate a dynamic plot for the case of mixture of crops, single simulation version and each variable in different plot

plot_dynamic_mixture_overlap: Generate a dynamic plot for the case of mixture of crops, single simulation version and several variables in same plot (overlap)

plot_dynamic_versions: Generate a dynamic plot for the case of sole crops, all simulation versions in same plot and each variable in a plot

plot_dynamic_overlap: Generate a dynamic plot for the case of sole crops, single simulation version and several variables in same plot (overlap)

plot_dynamic_mixture_versions_overlap: not implemented (seems not useful because too complex to analyze)

plot_dynamic_versions_overlap: Generate a dynamic plot for the case of sole crops, all simulation versions and several variables in same plot (overlap)

plot_dynamic_mixture_versions: Generate a dynamic plot for the case of mixture of crops, all simulation versions in same plot and each variable in a plot