Specific functions to generate scatter plots

Generate scatter plots for the different cases handled in CroPlotR (plant mixture, plot of residuals, plot several simulation results on same graph, ...) as specified by the different arguments.

Compute axis bounds (+/-0.05 added to the min/max of the data).

Make axis square

Ensure that the Y axis includes zero when all values in a facet are strictly positive or strictly negative.

Add error bars on observed values in given scatterplot

compute_axis_bounds(df_data, reference_var, y_var_type, is_obs_sd)

make_axis_square(df_data, reference_var, y_var_type, is_obs_sd, p)

force_y_axis(df_data, reference_var, y_var_type, is_obs_sd, p)

add_obs_error_bars(p, colour_factor = NULL)

plot_scat_mixture_allsit(
  df_data,
  sit,
  select_scat,
  shape_sit,
  reference_var,
  is_obs_sd,
  title = NULL
)

plot_scat_mixture_versions(
  df_data,
  sit,
  select_scat,
  shape_sit,
  reference_var,
  is_obs_sd,
  title = NULL
)

plot_scat_allsit(
  df_data,
  sit,
  select_scat,
  shape_sit,
  reference_var,
  is_obs_sd,
  title = NULL,
  has_distinct_situations = FALSE,
  one_version = FALSE,
  mixture = FALSE
)

plot_scat_versions_per_sit(
  df_data,
  sit,
  select_scat,
  shape_sit,
  reference_var,
  is_obs_sd,
  title = NULL
)

plot_scat_versions_allsit(
  df_data,
  sit,
  select_scat,
  shape_sit,
  reference_var,
  is_obs_sd,
  title = NULL
)

Arguments

df_data

A named list of data frame including the data to plot (one df per situation, or only one df if sit==all_situations)

y_var_type

type of variable to plot ("Simulated" or "Residuals")

is_obs_sd

TRUE if error standard deviation of observations is provided

p

A ggplot to modify`

colour_factor

The factor to use for colouring the error bars

sit

The name of the situation to plot (or all_situations)

has_distinct_situations

TRUE if the plot is for several situations

one_version

TRUE if the plot is for one version

mixture

TRUE if the plot is for a mixture of crops

Value

A ggplot object

A list of ggplot objects

List of x and y axis bounds (xaxis_min, xaxis_max, yaxis_min, yaxis_max)

The modified ggplot

The modified ggplot

The modified ggplot

Details

List of the different specific functions:

plot_scat_mixture_allsit: Generate a scatter plot for the case of mixture of crops, single simulation version and all_situations in same plot

plot_scat_allsit: Generate a scatter plot for the case of sole crops, single simulation version and all_situations in same plot