Format simulations and observations from CropR format to a format usable by CroplotR

Format simulations (and observations if any) for plotting. This function can be used as a template to include other models in CroPlotR.

format_cropr(
  sim,
  obs = NULL,
  obs_sd = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  all_situations = FALSE,
  successive = NULL,
  reference_var = NULL,
  verbose = TRUE
)

Arguments

sim
A simulation list of data.frames

obs
An observation list of data.frames

obs_sd
A list of observation standard deviation data.frames

type
The type of plot required, either "dynamic" or "scatter"

select_dyn
Which data to plot when type= "dynamic"? See details.

select_scat
Which data to plot when type= "scatter"? See details.

all_situations
Boolean (default = FALSE). If TRUE, plot all situations on the same graph. If TRUE, sim and obs are respectively an element of the first element and the second element of the output of cat_situations.

successive
A list of lists containing the situations to be represented as a contiguous sequence when type = "dynamic" (implies that the situations are correctly ordered).

reference_var
Variable selected on x-axis when type is scatter and select_scat is res. It is possible to select between observation and simulation of the reference variable.

verbose
Boolean. Print information during execution.

Value

A pre-formatted data.frame or NULL if the formatting is not possible (e.g. type="scatter" but no common variables in obs and sim).

Details

The select_dyn argument can be:

"sim" (the default): all variables with simulations outputs, and observations when there are some

"common": variables with simulations outputs and observations in common (used when type= "scatter" )

"obs": all variables with observations, and simulations outputs when there are some

"all": all variables with any observations or simulations outputs

The select_scat argument can be:

"sim" (the default): plots observations in X and simulations in Y.

"res": plots observations in X and residuals (observations-simulations) in Y.

Examples

workspace= system.file(file.path("extdata", "stics_example_1"), package = "CroPlotR")
situation= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))[1]
sim= SticsRFiles::get_daily_results(workspace = workspace, usm_name = situation)

#>            pIC_Wheat_Pea_2005-2006_N0            aIC_Wheat_Pea_2005-2006_N0 
#> "mod_spIC_Wheat_Pea_2005-2006_N0.sti" "mod_saIC_Wheat_Pea_2005-2006_N0.sti" 

obs= SticsRFiles::get_obs(workspace =  workspace, usm_name = situation)

#>       IC_Wheat_Pea_2005-2006_N0p       IC_Wheat_Pea_2005-2006_N0a 
#> "IC_Wheat_Pea_2005-2006_N0p.obs" "IC_Wheat_Pea_2005-2006_N0a.obs" 

formated_df= format_cropr(sim$`IC_Wheat_Pea_2005-2006_N0`,obs$`IC_Wheat_Pea_2005-2006_N0`)
options(max.print= 100)
formated_df

#>          Date   Plant Dominance variable Simulated Observed
#> 1  2005-09-26 plant_1 Principal    lai_n         0       NA
#> 2  2005-09-27 plant_1 Principal    lai_n         0       NA
#> 3  2005-09-28 plant_1 Principal    lai_n         0       NA
#> 4  2005-09-29 plant_1 Principal    lai_n         0       NA
#> 5  2005-09-30 plant_1 Principal    lai_n         0       NA
#> 6  2005-10-01 plant_1 Principal    lai_n         0       NA
#> 7  2005-10-02 plant_1 Principal    lai_n         0       NA
#> 8  2005-10-03 plant_1 Principal    lai_n         0       NA
#> 9  2005-10-04 plant_1 Principal    lai_n         0       NA
#> 10 2005-10-05 plant_1 Principal    lai_n         0       NA
#> 11 2005-10-06 plant_1 Principal    lai_n         0       NA
#> 12 2005-10-07 plant_1 Principal    lai_n         0       NA
#> 13 2005-10-08 plant_1 Principal    lai_n         0       NA
#> 14 2005-10-09 plant_1 Principal    lai_n         0       NA
#> 15 2005-10-10 plant_1 Principal    lai_n         0       NA
#> 16 2005-10-11 plant_1 Principal    lai_n         0       NA
#>  [ reached 'max' / getOption("max.print") -- omitted 1224 rows ]