Format a STICS simulation (and observation)

Format a STICS simulation (and observation if any) for plotting. This function can be used as a template to include other models in CroPlotR.

format_stics(
  sim,
  obs = NULL,
  obs_sd = NULL,
  type = c("dynamic", "scatter"),
  select_dyn = c("sim", "common", "obs", "all"),
  select_scat = c("sim", "res"),
  all_situations = FALSE,
  successive = NULL,
  reference_var = NULL
)

Arguments

sim
A simulation list of data.frames

obs
An observation list of data.frames

obs_sd
A list of observation standard deviation data.frames

type
The type of plot required, either "dynamic" or "scatter"

select_dyn
Which data to plot when type= "dynamic"? See details.

select_scat
Which data to plot when type= "scatter"? See details.

all_situations
Boolean (default = FALSE). If TRUE, plot all situations on the same graph. If TRUE, sim and obs are respectively an element of the first element and the second element of the output of cat_situations.

successive
A list of lists containing the situations to be represented as a contiguous sequence when type = "dynamic" (implies that the situations are correctly ordered).

reference_var
Variable selected on x-axis when type is scatter and select_scat is res. It is possible to select between observation and simulation of the reference variable.

Value

A pre-formatted data.frame or NULL if the formatting is not possible (e.g. type="scatter" but no common variables in obs and sim).

Details

The select_dyn argument can be:

"sim" (the default): all variables with simulations outputs, and observations when there are some

"common": variables with simulations outputs and observations in common (used when type= "scatter" )

"obs": all variables with observations, and simulations outputs when there are some

"all": all variables with any observations or simulations outputs

The select_scat argument can be:

"sim" (the default): plots observations in X and simulations in Y.

"res": plots observations in X and residuals (observations-simulations) in Y.

Examples

workspace= system.file(file.path("extdata", "stics_example_1"), package = "CroPlotR")
situation= SticsRFiles::get_usms_list(usm_path = file.path(workspace,"usms.xml"))[1]
sim= SticsRFiles::get_daily_results(workspace = workspace, usm_name = situation)
obs= SticsRFiles::get_obs(workspace =  workspace, usm_name = situation)
formated_df= format_stics(sim$`IC_Wheat_Pea_2005-2006_N0`,obs$`IC_Wheat_Pea_2005-2006_N0`)
options(max.print= 100)
formated_df

#>          Date Plant Dominance variable Simulated
#> 1  2005-09-26   ble Principal    lai_n         0
#> 2  2005-09-27   ble Principal    lai_n         0
#> 3  2005-09-28   ble Principal    lai_n         0
#> 4  2005-09-29   ble Principal    lai_n         0
#> 5  2005-09-30   ble Principal    lai_n         0
#> 6  2005-10-01   ble Principal    lai_n         0
#> 7  2005-10-02   ble Principal    lai_n         0
#> 8  2005-10-03   ble Principal    lai_n         0
#> 9  2005-10-04   ble Principal    lai_n         0
#> 10 2005-10-05   ble Principal    lai_n         0
#> 11 2005-10-06   ble Principal    lai_n         0
#> 12 2005-10-07   ble Principal    lai_n         0
#> 13 2005-10-08   ble Principal    lai_n         0
#> 14 2005-10-09   ble Principal    lai_n         0
#> 15 2005-10-10   ble Principal    lai_n         0
#> 16 2005-10-11   ble Principal    lai_n         0
#> 17 2005-10-12   ble Principal    lai_n         0
#> 18 2005-10-13   ble Principal    lai_n         0
#> 19 2005-10-14   ble Principal    lai_n         0
#> 20 2005-10-15   ble Principal    lai_n         0
#>  [ reached 'max' / getOption("max.print") -- omitted 1220 rows ]