Model quality assessment

Provide several metrics to assess the quality of the predictions of a model (see note) against observations.

r_means(sim, obs, na.rm = T)

R2(sim, obs, na.action = stats::na.omit)

SS_res(sim, obs, na.rm = T)

Inter(sim, obs, na.action = stats::na.omit)

Slope(sim, obs, na.action = stats::na.omit)

RMSE(sim, obs, na.rm = T)

RMSEs(sim, obs, na.rm = T)

RMSEu(sim, obs, na.rm = T)

nRMSE(sim, obs, na.rm = T)

rRMSE(sim, obs, na.rm = T)

rRMSEs(sim, obs, na.rm = T)

rRMSEu(sim, obs, na.rm = T)

pRMSEs(sim, obs, na.rm = T)

pRMSEu(sim, obs, na.rm = T)

SDSD(sim, obs, na.rm = T)

LCS(sim, obs, na.rm = T)

rbias(sim, obs, na.rm = T)

rSDSD(sim, obs, na.rm = T)

rLCS(sim, obs, na.rm = T)

MAE(sim, obs, na.rm = T)

ABS(sim, obs, na.rm = T)

MSE(sim, obs, na.rm = T)

EF(sim, obs, na.rm = T)

NSE(sim, obs, na.rm = T)

Bias(sim, obs, na.rm = T)

MAPE(sim, obs, na.rm = T)

FVU(sim, obs, na.rm = T)

RME(sim, obs, na.rm = T)

tSTUD(sim, obs, na.rm = T)

tLimit(sim, obs, risk = 0.05, na.rm = T)

Decision(sim, obs, risk = 0.05, na.rm = T)

Arguments

sim
Simulated values

obs
Observed values

na.rm
Boolean. Remove NA values if TRUE (default)

na.action
A function which indicates what should happen when the data contain NAs.

risk
Risk of the statistical test

Value

A statistic depending on the function used.

Details

The statistics for model quality can differ between sources. Here is a short description of each statistic and its equation (see html version for LATEX):

r_means(): Ratio between mean simulated values and mean observed values (%), computed as : $$r_means = \frac{100*\frac{\sum_1^n(\hat{y_i})}{n}} {\frac{\sum_1^n(y_i)}{n}}$$

R2(): coefficient of determination, computed using stats::lm() on obs~sim.

SS_res(): residual sum of squares (see notes).

Inter(): Intercept of regression line, computed using stats::lm() on sim~obs.

Slope(): Slope of regression line, computed using stats::lm() on sim~obs.

RMSE(): Root Mean Squared Error, computed as $$RMSE = \sqrt{\frac{\sum_1^n(\hat{y_i}-y_i)^2}{n}}$$

RMSEs(): Systematic Root Mean Squared Error, computed as $$RMSEs = \sqrt{\frac{\sum_1^n(\sim{y_i}-y_i)^2}{n}}$$ RMSEs = sqrt(mean((fitted.values(lm(formula=sim~obs))-obs)^2)

RMSEu(): Unsystematic Root Mean Squared Error, computed as $$RMSEu = \sqrt{\frac{\sum_1^n(\sim{y_i}-\hat{y_i})^2}{n}}$$ RMSEu = sqrt(mean((fitted.values(lm(formula=sim~obs))-sim)^2)

NSE(): Nash-Sutcliffe Efficiency, alias of EF, provided for user convenience.

nRMSE(): Normalized Root Mean Squared Error, also denoted as CV(RMSE), and computed as: $$nRMSE = \frac{RMSE}{\hat{y}}\cdot100$$

rRMSE(): Relative Root Mean Squared Error, computed as: $$rRMSE = \frac{RMSE}{\hat{y}}$$

rRMSEs(): Relative Systematic Root Mean Squared Error, computed as $$rRMSEs = \frac{RMSEs}{\hat{y}}$$

rRMSEu(): Relative Unsystematic Root Mean Squared Error, computed as $$rRMSEu = \frac{RMSEu}{\hat{y}}$$

pRMSEs(): Proportional Systematic Root Mean Squared Error, computed as: $$pRMSEs = \frac{RMSEs^2}{RMSE^2}$$

pRMSEu(): Proportional Unsystematic Root Mean Squared Error, computed as: $$pRMSEu = \frac{RMSEu^2}{RMSE^2}$$

SDSD(): Difference between sd_obs and sd_sim squared, computed as: $$SDSD = (sd_obs-sd_sim)^2$$

LCS(): Correlation between observed and simulated values, computed as: $$LCS = $$

rbias(): Relative bias squared, computed as: $$rbiase = \frac{Bias^2}{\hat{y}^2}$$

rSDSD(): Relative difference between sd_obs and sd_sim squared, computed as: $$rSDSD = \frac{SDSD}{\hat{y}^2}$$

rLCS(): Relative correlation between observed and simulated values, computed as: $$rLCS = \frac{LCS}{\hat{y}^2}$$

MAE(): Mean Absolute Error, computed as: $$MAE = \frac{\sum_1^n(\left|\hat{y_i}-y_i\right|)}{n}$$

ABS(): Mean Absolute Bias, which is an alias of MAE()

FVU(): Fraction of variance unexplained, computed as: $$FVU = \frac{SS_{res}}{SS_{tot}}$$

MSE(): Mean squared Error, computed as: $$MSE = \frac{1}{n}\sum_{i=1}^n(Y_i-\hat{Y_i})^2$$

EF(): Model efficiency, also called Nash-Sutcliffe efficiency (NSE). This statistic is related to the FVU as \(EF= 1-FVU\). It is also related to the \(R^2\) because they share the same equation, except SStot is applied relative to the identity function (i.e. 1:1 line) instead of the regression line. It is computed as: $$EF = 1-\frac{SS_{res}}{SS_{tot}}$$

Bias(): Modelling bias, simply computed as: $$Bias = \frac{\sum_1^n(\hat{y_i}-y_i)}{n}$$

MAPE(): Mean Absolute Percent Error, computed as: $$MAPE = \frac{\sum_1^n(\frac{\left|\hat{y_i}-y_i\right|}{y_i})}{n}$$

RME(): Relative mean error (\ $$RME = \frac{\sum_1^n(\frac{\hat{y_i}-y_i}{y_i})}{n}$$

tSTUD(): T student test of the mean difference, computed as: $$tSTUD = \frac{Bias}{\sqrt(\frac{var(M)}{n_obs})}$$

tLimit(): T student threshold, computed using qt(): $$tLimit = qt(1-\frac{\alpha}{2},df=length(obs)-1)$$

Decision(): Decision of the t student test of the mean difference, computed as: $$Decision = tSTUD + tLimit \gt 0$$

Note

\(SS_{res}\) is the residual sum of squares and \(SS_{tot}\) the total sum of squares. They are computed as: $$SS_{res} = \sum_{i=1}^n (y_i - \hat{y_i})^2$$ $$SS_{tot} = \sum_{i=1}^{n}\left(y_{i}-\bar{y}\right)^2$$ Also, it should be noted that \(y_i\) refers to the observed values and \(\hat{y_i}\) to the predicted values, \(\bar{y}\) to the mean value of observations and \(\sim{y_i}\) to values predicted by linear regression.

Examples

if (FALSE) {
sim= rnorm(n = 5,mean = 1,sd = 1)
obs= rnorm(n = 5,mean = 1,sd = 1)
RMSE(sim,obs)
}