Generic statistics for all situations

Compute statistics for simulations of one or several situations with its observations, eventually grouped by a model version (or any group actually).

statistics_situations(
  ...,
  obs = NULL,
  stat = "all",
  all_situations = TRUE,
  all_plants = TRUE,
  verbose = TRUE,
  formater
)

Arguments

...
Simulation outputs (each element= model version), each being a named list of data.frame for each situation. See examples.

obs
A list (each element= situation) of observations data.frames (named by situation)

stat
A character vector of required statistics, "all" for all, or any of predictor_assessment().

all_situations
Boolean (default = TRUE). If TRUE, computes statistics for all situations.

all_plants
Boolean (default = TRUE). If TRUE, computes statistics for all plants (when applicable).

verbose
Boolean. Print information during execution.

formater
The function used to format the models outputs and observations in a standard way. You can design your own function that format one situation and provide it here (see statistics() and format_cropr() for more information).

Value

A tibble::as_tibble() with statistics grouped by group (i.e. model version) and situation

See also