Generic statistics for all situations

Compute statistics for simulations of one or several situations with its observations, eventually grouped by a model version (or any group actually).

statistics_situations(..., obs = NULL, stat = "all", verbose = TRUE, formater)

Arguments

...
Simulation outputs (each element= model version), each being a named list of data.frame for each situation. See examples.

obs
A list (each element= situation) of observations data.frames (named by situation)

stat
A character vector of required statistics, "all" for all, or any of predictor_assessment().

verbose
Boolean. Print informations during execution.

formater
The function used to format the models outputs and observations in a standard way. You can design your own function that format one situation and provide it here (see statistics() and format_stics() for more information).

Value

A dplyr::tibble() with statistics grouped by group (i.e. model version) and situation

See also